requirements.txt
langchain
langgraph
langchain-community
streamlit
tiktoken
faiss-cpu
python-dotenv

